+++ src/components/server/Oauth2AuthorizationServer.php
@@ -7,8 +7,19 @@ use rhertogh\Yii2Oauth2Server\interfaces\components\server\Oauth2AuthorizationSe
 
 class Oauth2AuthorizationServer extends AuthorizationServer implements Oauth2AuthorizationServerInterface
 {
+    /**
+     * @inheritdoc
+     */
     public function getEnabledGrantTypes()
     {
         return $this->enabledGrantTypes;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getGrantTypesAccessTokenTTLs()
+    {
+        return $this->grantTypeAccessTokenTTL;
     }
 }