+++ src/interfaces/components/factories/grants/base/Oauth2GrantTypeFactoryInterface.php
@@ -3,8 +3,9 @@
 namespace rhertogh\Yii2Oauth2Server\interfaces\components\factories\grants\base;
 
 use League\OAuth2\Server\Grant\GrantTypeInterface;
+use rhertogh\Yii2Oauth2Server\interfaces\components\common\DefaultAccessTokenTtlInterface;
 
-interface Oauth2GrantTypeFactoryInterface
+interface Oauth2GrantTypeFactoryInterface extends DefaultAccessTokenTtlInterface
 {
     /**
      * Get the server grant type.